+++ lib/onlyoffice_bugzilla_helper/comments.rb
@@ -8,5 +8,17 @@ module OnlyofficeBugzillaHelper
       end
       JSON.parse(res.body)['bugs'][bug_id.to_s]['comments']
     end
+
+    # Add comment to bug
+    # @param bug_id [Integer] id of bug
+    # @param comment [String] comment to add
+    def add_comment(bug_id, comment)
+      req = Net::HTTP::Post.new(bug_url(bug_id, '/comment'))
+      req.body = { comment: comment }.to_json
+      req.add_field('Content-Type', 'text/plain')
+      connection = Net::HTTP.new(@uri.host, @uri.port)
+      connection.use_ssl = use_ssl?
+      connection.start { |http| http.request(req) }
+    end
   end
 end