+++ spec/onlyoffice_bugzilla_helper/bugzilla_helper_comments_spec.rb
@@ -7,4 +7,11 @@ RSpec.describe OnlyofficeBugzillaHelper::BugzillaHelper, '#comments' do
     comments = bugzilla.comments(BUG_ID_TEST_STATUS)
     expect(comments.first['text']).to include('Сценарий')
   end
+
+  it 'Add comment to bug' do
+    comment = Faker::StarWars.quote
+    bugzilla.add_comment(BUG_ID_TEST_STATUS, comment)
+    comments = bugzilla.comments(BUG_ID_TEST_STATUS)
+    expect(comments.last['text']).to eq(comment)
+  end
 end